requirements.txt
Flask==2.1.1
numpy==1.22.3
Pillow==9.1.0
Werkzeug==2.1.1
gunicorn==20.1.0
